Remove console (getty) upstart script
Upstart package (in meta-webos) already creates upstart scripts
for ttyS0 (tty-serial, based on SERIAL_CONSOLE in <machine>.conf)
and tty1 (based on SYSVINIT_ENABLED_GETTYS in <machine>.conf or
upstart .bb file).

The console upstart script is a legacy webOS file, and actually
conflicts with the upstart getty scripts by launching a getty on
/dev/console. When two are active on the same device, neither works.

If a getty is needed on a different device for a given machine, it
should be defined in SERIAL_CONSOLE or SYSVINIT_ENABLED_GETTYS in
the <machine.conf> as noted above.

Open-webOS-DCO-1.0-Signed-off-by: Steve Lemke <steve.lemke@palm.com>

Change-Id: I25bc4101ffac9d27b977c4eb28198141b6c33030
Reviewed-on: http://g2g.palm.com/926
Reviewed-by: Build Verification
Reviewed-by: Steve Lemke <steve.lemke@palm.com>
Tested-by: Steve Lemke <steve.lemke@palm.com>
Reviewed-by: Keith Derrick <keith.derrick@palm.com>
Reviewed-by: Herb Kuta <herb.kuta@palm.com>