Clean up wpa-supplicant for qemux
:Release Notes:
In the previous fix I missed a harmless but ugly bug in
wpa-supplicant.conf that caused it to fail 10 times rapidly on qemux86
instead of the clean exit I expected.

:Detailed Notes:
While the Upstart Cookbook states that 'exit 0' is a "good" exit and
will not cause a 'respawn' job to respawn, actual behavior is different
and a trick is needed to get the wpa-supplicant job to cleanly not run
under qemux86.  Since this trick results in the wpa-supplicant Upstart
job never entering the starting or started states, we need to
explicitly start connman under qemux.

:Testing Performed:
GPVB build running an auto-reboot looping test on qemux under
VirtualBox

:QA Notes:

:Issues Addressed:
[CTO-640] Fine tune boot sequence

Open-webOS-DCO-1.0-Signed-off-by: Ed Chejlava <ed.chejlava@lge.com>

Change-Id: I29132c8dfe2c47c36ef94149c84c92f98f6d1b8d
Reviewed-on: https://g2g.lgsvl.com/6496
Reviewed-by: DCO Verification
Tested-by: Ed Chejlava <ed.chejlava@lge.com>
Reviewed-by: Ed Chejlava <ed.chejlava@lge.com>